Split 'utility' functionality.
