+++ src/utility/mod.rs
@@ -1,0 +1,5 @@
+
+pub mod parser;
+
+mod bitfield;
+pub use self::bitfield::Bitfield;
